+++ nautilus_core/indicators/src/momentum/cmo.rs
@@ -119,8 +119,13 @@ impl ChandeMomentumOscillator {
             self.initialized = true;
         }
         if self.initialized {
-            self.value = 100.0 * (self._average_gain.value() - self._average_loss.value())
-                / (self._average_gain.value() + self._average_loss.value());
+            let divisor = self._average_gain.value() + self._average_loss.value();
+            if divisor == 0.0 {
+                self.value = 0.0;
+            } else {
+                self.value =
+                    100.0 * (self._average_gain.value() - self._average_loss.value()) / divisor;
+            }
         }
         self._previous_close = close;
     }